+++ tests/tree.rs
@@ -7,17 +7,18 @@
 #![cfg(feature = "canon")]
 
 mod max_annotation;
-use dusk_plonk::circuit;
+
 use dusk_plonk::error::Error as PlonkError;
-use dusk_plonk::prelude::*;
 use dusk_poseidon::tree::{
     self, PoseidonAnnotation, PoseidonBranch, PoseidonTree,
 };
 use max_annotation::MockLeaf;
 use rand_core::{CryptoRng, OsRng, RngCore};
 
+use dusk_plonk::prelude::*;
+
 const DEPTH: usize = 17;
-const CAPACITY: usize = 1 << 15;
+const CAPACITY: usize = 15;
 type Tree = PoseidonTree<MockLeaf, PoseidonAnnotation, DEPTH>;
 
 struct MerkleOpeningCircuit {
@@ -41,10 +42,6 @@ impl MerkleOpeningCircuit {
 
         Self { branch }
     }
-
-    pub fn public_inputs(&self) -> Vec<PublicInputValue> {
-        vec![(*self.branch.root()).into()]
-    }
 }
 
 impl Circuit for MerkleOpeningCircuit {
@@ -52,48 +49,50 @@ impl Circuit for MerkleOpeningCircuit {
 
     fn gadget(
         &mut self,
-        composer: &mut StandardComposer,
+        composer: &mut TurboComposer,
     ) -> Result<(), PlonkError> {
-        let root = self.branch.root();
-        let root_p = tree::merkle_opening::<DEPTH>(composer, &self.branch);
+        use std::ops::Deref;
 
-        composer.constrain_to_constant(root_p, BlsScalar::zero(), Some(-root));
+        let leaf: BlsScalar = *self.branch.deref();
+        let leaf = composer.append_witness(leaf);
+
+        let root = self.branch.root();
+        let root = composer.append_witness(*root);
+
+        let root_p =
+            tree::merkle_opening::<DEPTH>(composer, &self.branch, leaf);
+
+        composer.assert_equal(root_p, root);
 
         Ok(())
     }
 
-    fn padded_circuit_size(&self) -> usize {
-        CAPACITY
+    fn public_inputs(&self) -> Vec<PublicInputValue> {
+        vec![]
+    }
+
+    fn padded_gates(&self) -> usize {
+        1 << CAPACITY
     }
 }
 
 #[test]
 fn tree_merkle_opening() {
-    let mut rng = OsRng;
-    let pp = PublicParameters::setup(CAPACITY, &mut rng).unwrap();
     let label = b"dusk-network";
+    let pp = PublicParameters::setup(1 << CAPACITY, &mut OsRng).unwrap();
 
     let mut tree = Tree::default();
-    let mut circuit = MerkleOpeningCircuit::random(&mut rng, &mut tree);
+
+    let mut circuit = MerkleOpeningCircuit::random(&mut OsRng, &mut tree);
     let (pk, vd) = circuit.compile(&pp).expect("Failed to compile circuit");
 
     let mut tree = Tree::default();
-    for _ in 0..13 {
-        let mut circuit = MerkleOpeningCircuit::random(&mut rng, &mut tree);
+    let mut circuit = MerkleOpeningCircuit::random(&mut OsRng, &mut tree);
 
-        let proof = circuit
-            .gen_proof(&pp, &pk, label)
-            .expect("Failed to generate proof");
-        let pi = circuit.public_inputs();
+    let proof = circuit
+        .prove(&pp, &pk, label)
+        .expect("Failed to generate proof");
 
-        circuit::verify_proof(
-            &pp,
-            vd.key(),
-            &proof,
-            pi.as_slice(),
-            vd.pi_pos(),
-            label,
-        )
+    MerkleOpeningCircuit::verify(&pp, &vd, &proof, &[], label)
         .expect("Proof verification failed");
-    }
 }