+++ CHANGELOG.md
@@ -7,6 +7,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Update `dusk-bls12_381` from `0.8` to `0.9`
+- Update `dusk-jubjub` from `0.10` to `0.11`
+- Update `dusk-hades` from `0.17.0-rc` to `0.18.0-rc`
+- Update `canonical` from `0.6` to `0.7`
+- Update `canonical_derive` from `0.6` to `0.7`
+- Update `microkelvin` from `0.14` to `0.15`
+- Update `nstack` from `0.13` to `0.14.0-rc`
+- Update `dusk-plonk` from `0.9` to `0.10`
+- Change merkle opening to constrain leaf [#162]
+- Export `sponge::truncated::hash` regardless of `alloc` feature [#167]
+- Remove useless `let` in `sponge::truncated`
+
+### Fixed
+
+- Fix module injection for `tree` and `cipher` modules
+
 ## [0.22.0] - 2021-07-27
 
 ### Changed
@@ -37,6 +55,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Change `PoseidonLeaf` getter methods to return refs [#143]
 
 ### Removed
+
 - Remove `anyhow` and `thiserror` from deps [#132]
 - Remove `PoseidonWalkableIterator` and `PoseidonWalkableAnnotation` [#125]
 - Remove `canon_host` feature checks from CI [#136]
@@ -278,6 +297,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Poseidon252 Sponge-hash impl with BulletProofs.
 - Variants of sponge for `Scalar` & `Gadget(Variable/LC)`.
 
+[#167]: https://github.com/dusk-network/poseidon252/issues/167
+[#162]: https://github.com/dusk-network/poseidon252/issues/162
 [#158]: https://github.com/dusk-network/poseidon252/issues/158
 [#153]: https://github.com/dusk-network/poseidon252/issues/153
 [#151]: https://github.com/dusk-network/poseidon252/issues/151